Issue #16, extracted main method to Main class, refactored MonopolyGame constructor.
