+++ src/test/java/ch/heig/gen/lab03/MonopolyGameTest.java
@@ -2,6 +2,6 @@ package ch.heig.gen.lab03;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-class MgameTest {
+class MonopolyGameTest {
 
 }
